See Description for changes
- Added a temporary bool i ngameinstance to check if it should spawn a
player in the the new level or not. This fix removes the excess
characters from new levels.
- Made the character text follow camera on tick
- Made "win" widget to be displayed upon winning game. Has functions in
topdowncharacter and just needs to implemented in each level/called upon
when match is done. This widget will display which player won, what item
they recieved and after a countdown of 10 seconds it will return
everyone to the lobby.